refactor(rt): module rearrangements and documentation fixes
Rearrangements:
- Add feature 'k510', add k510 SoC module. Only one of the stacks from each SoC modules should be compiled.
- Add fallback stack and stack_size variables
- Split 'soc!' macro into 'macro' module
- Split start instructions into 'cpu_c908' module; reserve other 'cpu_*' modules
- Reserve space for K210 and K510 chips
- Reformat 'xtask/src/generate/image.rs' in Rust 2024 rustfmt conventions

Documentation fixes:
- Add a line at the end of each file
- Adjust format for Cargo.toml files
- Symbol '[' and ']' might be used as markdown hyperlinks; if we use as regular brackets we should use '\[' and '\]'
