+++ kendryte-rt/src/macros.rs
@@ -1,0 +1,37 @@
+macro_rules! soc {
+    (
+        $(
+            $(#[$doc:meta])*
+            pub struct $Ty:ident => $paddr:expr, $DerefTy:ty;
+        )+
+    ) => {
+        $(
+            $(#[$doc])*
+            #[allow(non_camel_case_types)]
+            pub struct $Ty(());
+
+            impl $Ty {
+                #[inline]
+                pub const fn ptr() -> *const $DerefTy {
+                    $paddr as *const $DerefTy
+                }
+            }
+
+            impl core::ops::Deref for $Ty {
+                type Target = $DerefTy;
+
+                #[inline(always)]
+                fn deref(&self) -> &'static Self::Target {
+                    unsafe { &*Self::ptr() }
+                }
+            }
+
+            impl core::convert::AsRef<$DerefTy> for $Ty {
+                #[inline(always)]
+                fn as_ref(&self) -> &'static $DerefTy {
+                    unsafe { &*Self::ptr() }
+                }
+            }
+        )+
+    };
+}
